converted datasource obects to JSON files
